internal/relpath: use os.Getwd on wasip2
